REST service finished

+++ src/main/java/net/psv73/insurance/repository/InsuranceRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface InsuranceRepository extends CrudRepository<Insurance, Integer> {
+public interface InsuranceRepository extends CrudRepository<Insurance, Long> {
 
     List<Insurance> findAll();
